+++ Soomla/Assets/Plugins/Soomla/Store/IOSImpl/VirtualGoodsStorageIOS.cs
@@ -22,7 +22,7 @@ namespace Soomla.Store
 	/// abstract <c>VirtualGoodsStorage</c> for Android.
 	/// </summary>
 	public class VirtualGoodsStorageIOS : VirtualGoodsStorage {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_IOS && !UNITY_EDITOR
 
 		/// Functions that call iOS-store functions.
 		[DllImport ("__Internal")]