+++ Soomla/Assets/Plugins/Soomla/Store/data/StoreInfo.cs
@@ -193,12 +193,37 @@ namespace Soomla.Store
 		}
 
 		/// <summary>
-		/// Replaces the given virtual item, and then saves the store's metadata.
+		/// Replaces the given virtual item, and then saves the store's metadata
+		/// (if requested).
 		/// </summary>
 		/// <param name="virtualItem">the virtual item to replace.</param>
-		public static void Save(VirtualItem virtualItem) {
+		/// <param name="saveToDB">should the virtual item be persisted to local DB</param>
+		public static void Save(VirtualItem virtualItem, bool saveToDB = true) {
 			replaceVirtualItem(virtualItem);
-			Save();
+
+			if (saveToDB) {
+				Save();
+			}
+		}
+
+		/// <summary>
+		/// Replaces the given virtual item, and then saves the store's metadata
+		/// (if requested).
+		/// </summary>
+		/// <param name="virtualItem">the virtual item to replace.</param>
+		/// <param name="saveToDB">should the virtual item be persisted to local DB</param>
+		public static void Save(List<VirtualItem> virtualItems, bool saveToDB = true) {
+			if ((virtualItems == null) && (virtualItems.Count == 0)) {
+				return;
+			}
+
+			foreach(VirtualItem virtualItem in virtualItems) {
+				replaceVirtualItem(virtualItem);
+			}
+
+			if (saveToDB) {
+				Save();
+			}
 		}
 
 
@@ -384,7 +409,7 @@ namespace Soomla.Store
 					List<UpgradeVG> upgrades;
 					if (!GoodsUpgrades.TryGetValue (((UpgradeVG)vi).GoodItemId, out upgrades)) {
 						upgrades = new List<UpgradeVG> ();
-						GoodsUpgrades.AddOrUpdate(((UpgradeVG)vi).GoodItemId, upgrades);
+						GoodsUpgrades.Add(((UpgradeVG)vi).GoodItemId, upgrades);
 					}
 					upgrades.Add ((UpgradeVG)vi);
 				}
@@ -441,10 +466,10 @@ namespace Soomla.Store
 				VirtualGood vg = (VirtualGood)virtualItem;
 				
 				if (vg is UpgradeVG) {
-					List<UpgradeVG> upgrades = GoodsUpgrades[((UpgradeVG) vg).GoodItemId];
-					if (upgrades == null) {
+					List<UpgradeVG> upgrades;
+					if (!GoodsUpgrades.TryGetValue (((UpgradeVG) vg).GoodItemId, out upgrades)) {
 						upgrades = new List<UpgradeVG>();
-						GoodsUpgrades.AddOrUpdate(((UpgradeVG) vg).ItemId, upgrades);
+						GoodsUpgrades.Add(((UpgradeVG) vg).ItemId, upgrades);
 					}
 					upgrades.Add((UpgradeVG) vg);
 				}