+++ Soomla/Assets/Plugins/Soomla/Store/purchaseTypes/PurchaseWithVirtualItem.cs
@@ -64,37 +64,62 @@ namespace Soomla.Store
 			SoomlaUtils.LogDebug("SOOMLA PurchaseWithVirtualItem", "Trying to buy a " + AssociatedItem.Name + " with "
 			                     + Amount + " pieces of " + TargetItemId);
 
-			VirtualItem item = null;
-			try {
-				item = StoreInfo.GetItemByItemId(TargetItemId);
-			} catch (VirtualItemNotFoundException) {
-				SoomlaUtils.LogError(TAG, "Target virtual item doesn't exist !");
+			VirtualItem item = getTargetVirtualItem ();
+			if (item == null) {
 				return;
 			}
 
+
 			JSONObject eventJSON = new JSONObject();
 			eventJSON.AddField("itemId", AssociatedItem.ItemId);
-			StoreEvents.Instance.onItemPurchaseStarted(eventJSON.print());
+			StoreEvents.Instance.onItemPurchaseStarted(eventJSON.print(), true);
 
-			int balance = item.GetBalance();
-			if (item is VirtualCurrency) {
-				balance = VirtualCurrencyStorage.GetBalance(item);
-			} else {
-				balance = VirtualGoodsStorage.GetBalance(item);
-			}
-
-			if (balance < Amount){
-				throw new InsufficientFundsException(TargetItemId);
+			if (!checkTargetBalance (item)) {
+				throw new InsufficientFundsException (TargetItemId);
 			}
 
 			item.Take(Amount);
 
 			AssociatedItem.Give(1);
 
-			eventJSON = new JSONObject();
-			eventJSON.AddField("itemId", AssociatedItem.ItemId);
-			eventJSON.AddField("payload", payload);
-			StoreEvents.Instance.onItemPurchased(eventJSON.print());
+			// We have to make sure the ItemPurchased event will be fired AFTER the balance/currency-changed events.
+			StoreEvents.Instance.RunLater(() => {
+				eventJSON = new JSONObject();
+				eventJSON.AddField("itemId", AssociatedItem.ItemId);
+				eventJSON.AddField("payload", payload);
+				StoreEvents.Instance.onItemPurchased(eventJSON.print(), true);
+			});
+		}
+		
+		/// <summary>
+		/// Checks if there is enough funds to afford the <code>PurchasableVirtualItem</code>.
+		/// Implementation in subclasses will be according to specific type of purchase.
+		/// </summary>
+		/// <returns>True if there are enough funds to afford the virtual item with the given item id </returns>
+		public override bool CanAfford() {
+			SoomlaUtils.LogDebug("SOOMLA PurchaseWithVirtualItem", "Checking affordability of " + AssociatedItem.Name + " with "
+			                     + Amount + " pieces of " + TargetItemId);
+
+			VirtualItem targetItem = getTargetVirtualItem();
+			return checkTargetBalance(targetItem);
+		}
+
+		private VirtualItem getTargetVirtualItem ()
+		{
+			VirtualItem item = null;
+			try {
+				item = StoreInfo.GetItemByItemId (TargetItemId);
+			}
+			catch (VirtualItemNotFoundException) {
+				SoomlaUtils.LogError (TAG, "Target virtual item doesn't exist !");
+			}
+			return item;
+		}
+
+		private bool checkTargetBalance (VirtualItem item)
+		{
+			int balance = item.GetBalance ();
+			return balance >= Amount;
 		}
 	}
 }