+++ Soomla/Assets/Soomla/Editor/build-tools/Soomla_StoreRunner.py
@@ -1,3 +1,17 @@
+# Copyright (C) 2012-2014 Soomla Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 #!/usr/bin/python
 
 from mod_pbxproj import *
@@ -7,7 +21,6 @@ import sys
 
 frameworks = [
   'System/Library/Frameworks/Security.framework',
-  'usr/lib/libsqlite3.0.dylib',
   'System/Library/Frameworks/StoreKit.framework'
 ]
 
@@ -19,10 +32,10 @@ pbx_file_path = sys.argv[1] + '/Unity-iPhone.xcodeproj/project.pbxproj'
 pbx_object = XcodeProject.Load(pbx_file_path)
 
 for framework in frameworks:
-  pbx_object.add_file(framework, tree='SDKROOT')
+  pbx_object.add_file_if_doesnt_exist(framework, tree='SDKROOT')
 
 for framework in weak_frameworks:
-  pbx_object.add_file(framework, tree='SDKROOT', weak=True)
+  pbx_object.add_file_if_doesnt_exist(framework, tree='SDKROOT', weak=True)
 
 pbx_object.add_other_ldflags('-ObjC')